pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.2",
    "uvicorn>=0.35.0",
]
